Intermittent failure in testRedeployOnJavaChangeWithCustomLauncher
Fixes #534

In this test, Main is a simple Java class that does not extend Launcher.
When redeployment occurs, the process does not stop cleanly and the server socket may still be bound when the new process starts.

Adding a shutdown hook should guarantee that the process stops cleanly and the new process can bind the server stockets when it starts.