Support both Vert.x 5 and Vert.x 4 projects (#593)
* Removed redundant class FileFilterMain

* Remove Vert.x from tests dependencies

Make sure we can run tests without depending on a particular Vert.x version.

* Remove support for dependencies argument in setup mojo

Closes #594

* Remove start and stop mojos

Closes #595

* Minor fixes to _vertx-initialize.adoc

* Minor fixes to docs

* Remove redundant fields and comment in AbstractVertxMojo

* Replace leftover hardcoded 3.9 vert.x version in invoker tests

* Implement redeployment in the plugin

This commit introduces an adapter that helps create the forked Vert.x process (there are some differences between Vert.x 4 and Vert.x 5 in this regard).

The debug mojo is merged into the run mojo. This way it is possible to make changes and reconnect immediately with a remote debugger, instead of interrupting a Maven build running vertx:run to start another one running vertx:debug.

Also, disabling the blocked thread checker is now a separate option instead of being implied by debug options.

* vertx-maven-plugin.properties should use configured version

The compiler and resources plugin versions are used in MojoUtils and ultimately in RunMojo to create default build tasks.

The versions used shouldn't be hardcoded. Instead, the plugin should read the versions configured in the POM file.

* Some test simplifications

* Manage test temp files and folders with junit rule

* Setup mojo should support Vert.x 5 projects

When generating a Vert.x 5 project, the mojo should add the vertx-application-launcher dependency in addition to the vertx-core.

Also, some updates have been made to use defaults instead of prompting when the Maven execution is not interactive.

* Support packaging, setup and deployment of Vert.x 5 applications

IT run with JDK11+, as it is the minimum Java version required by Vert.x 5.