Document how the plugin detects if a custom main class is a launcher
If the user provides a custom main class, the plugin must determine if it is a launcher.
For example, if it is a launcher, the run goal will add several arguments to the command line (like verticle name and number of instances).

This new doc section clarifies how the plugin decides whether the class is a launcher.